Refactor VMBoard component and remove EmptySectionRow
- Updated VMBoard component to support rendering an empty placeholder when no boards are present, enhancing user experience in the sidebar.
- Introduced a new VMBoardEmpty component to handle the display of empty states, improving code organization and clarity.
- Removed the obsolete EmptySectionRow component from the codebase, streamlining the component structure and reducing redundancy.
- Adjusted related components to utilize the new VMBoard structure, ensuring consistent behavior across the sidebar.
